If we're receiving packets from a PF_PACKET socket, check whether the interface index of the interface for the packet is the interface index of the loopback interface and, if it is, check if the packet is an outgoing packet; if so, ignore it, as we'll also be seeing that packet as a received packet.
If we don't handle the arphrd type of an interface, and fall back on
cooked mode, report the arphrd type, so we know what type we should
consider supporting (if that type can't be supported well, e.g.  if you
don't get any link-layer header, as happens with PPP, we'd be silent).