On AIX, if binding to the SAP 1537 fails, try binding to 0; Token Ring devices appear to reject attempts to bind to 1537, perhaps because Token Ring devices use SAPs rather than Ethertypes and 1537 isn't a valid SAP value.
Try to supply a string rather than a numerical value for various DLPI
errors, and to supply a string rather than a numerical value for
unexpected DLPI primitives.

Cast the argument to <ctype.h> macros to "unsigned char", to eliminate
GCC warnings and to keep the macros from referring outside an array when
handed bytes with the 8th bit set.