Patch, based on a patch from Pavel Kankovsky <kan@dcit.cz>, to flush all packets queued up on the socket when we set a kernel filter on the socket, so that if there are any queue-up packets that wouldn't have passed the new filter, we don't see them.  (Some other packet capture mechanisms do this automatically; this prevents tcpdump, for example, from showing or saving, when run with a filter, some packets that wouldn't have passed the filter.)
XXX - do we have to do this on any other platforms?

Choose whether to compile in the code to modify filter programs for use
in the kernel, and to flush queued-up packet and set a kernel filter, on
whether SO_ATTACH_FILTER is defined (i.e., on whether we have kernel
filter support in our build environment), rather than on whether
HAVE_PF_PACKET_SOCKETS is defined (i.e., on whether we have PF_PACKET
support in our build environment), as we choose whether to *use* that
code based on whether SO_ATTACH_FILTER is defined.