We don't actually use HAVE_FREEIFADDRS, so there's no point in checking whether we have "freeifaddrs()" (we don't check whether we have "getifaddrs()", and if we have "getifaddrs()" but not "freeifaddrs()", we're stuck with leaking memory).
Give the "any" device an instance number of INT_MAX, so it shows up
after all other non-loopback devices.

"getifaddrs()" sometimes appears to supply a destination address even
for non-point-to-point interfaces (it did so on a FreeBSD 4.1 system);
don't use the broadcast address it supplies if an interface isn't a
broadcast interface, and don't use the destination address it supplies
if an interface isn't a point-to-point interface.

If we had an error constructing the list of interfaces, don't attempt to
add the "any" device to the list.