Have "map_arphrd_to_dlt()" take a flag indicating whether it's OK to return DLT_LINUX_SLL or not, and, if that flag is false, for those interface types where we'd used DLT_LINUX_SLL, pick a DLT_ type that works as well as possible in raw mode, or fail.
Pass 1 as that flag if we're using a PF_PACKET socket; pass 0 as that
flag if we're using a PF_INET/SOCK_PACKET socket.

For PF_INET/SOCK_PACKET sockets, try to get the link-layer type and map
it to a DLT_ value *before* turning promiscuous mode on, so that we
don't try to put the interface into promiscuous mode unless we know we
can handle its link-layer type (and thus that we can use the interface).