When installing a filter:
	first check whether BIOCVERSION works;

	then, if so, check whether the version is OK for us;

	then, if so, try to install the filter - if that succeeds, free
	up any userland filter we had installed, if any, and filter in
	the kernel, otherwise fail;

	otherwise, if we couldn't get the version or the version check
	failed, install a userland filter.

That way, we don't install kernel filters unless we know the kernel's
BPF interpreter is OK for us (rather than checking the version *after*
we've already installed the filter), and don't have to worry about that
filter being in the kernel (possibly with an inappropriate interpreter)
if we're filtering in userland.

Clean up the "version isn't OK" message (put in a newline at the end,
and put in "tcpdump:" at the beginning).