Make "link[N:M]" refer to the 802.11 header for all 802.11 DLT_ values, including those with fixed-length radio headers (it already refers to the 802.11 header for radiotap).
Add a new "radio" keyword, to allow access to the radio header.  In
theory, something to allow testing for specific signal strengths, etc.
might be useful, but radiotap makes that difficult as the code can't
loop through the header looking for the signal strength field, the loop
has to be unrolled, and some of the other headers might not have
standardized the meaning of some of the fields, so we require the user
to construct such a filter themselves, for now.