Checking whether the amount of data returned by SIOCGIFCONF is less than the size of the buffer we handed to it is insufficient to determine whether we have the entire list of interfaces or not - if the amount of space left in the buffer after adding an entry is non-zero but less than the amount of space required by the next entry, the ioctl will stop before adding the next entry, and not necessarily return an error.
The only way to ensure that we got all the data is to pass a buffer
large enough that the amount of space in the buffer *not* filled in
is greater than the largest possible entry.

We assume that's "sizeof(ifreq.ifr_name)" plus 255, under the assumption
that no address is more than 255 bytes (on systems where the "sa_len"
field in a "struct sockaddr" is 1 byte, e.g. newer BSDs, that's the
case, and addresses are unlikely to be bigger than that in any case).