By default, don't include DEPLIBS in the libraries flags, as a dynamically-linked libpcap should have been linked with them (if there are any of them), so it shouldn't be necessary for a program or library to explicitly link with them if it links with libpcap.
Add a -static flag that includes DEPLIBS, as, on most if not all
platforms, static libraries can't be linked with dynamic libraries, so
programs would have to link with libraries on which libpcap depends if
it links statically with libpcap.

If both --cflags and --libs are given, print both sets of flags, on the
same line (as pkg-config does).