Fix a problem that shows up with a max packet count passed to pcap_dispatch().
If the last record read in by a read() from a BPF device isn't a
multiple of the alignment value for BPF_WORDALIGN(), we could increment
bp past ep; handle that case.

While we're at it, properly set p->bp and p->cc in the case where we
break out of the loop due to a pcap_breakloop() call.