Use config.status to generate pcap-config.
Rather than use a custom sed script to replace things in pcap-config,
use the already existing config.status script.  This keeps people from
having to update the Makefile.in every time they want to add a new var
to the file.

Also, clean up the pcap-config.in script a bit to set local vars at the
top of the file and then refer to those rather than replacing all things
through out.  This makes it easier to tweak on the fly as there is only
line to change rather than trying and figure out all the locations.

Reviewed-By: Guy Harris <guy@alum.mit.edu>