Give more details about the addresses in the address list.
Note that they might be IPv4 addresses, IPv6 addresses, or other
addresses of some unspecified sort, and that applications must therefore
check the sa_family field of the address before interpreting the
address.  Note also that IPv4 addresses are "struct sockaddr_in"s and
IPv6 addresses are "struct sockaddr_in6"s, so that people don't freak
out if they have to figure out the address size i order to pass it to
getnameinfo().  (And note that "struct sockaddr_in6" might not be
defined on crufty old OSes that don't support IPv6.  As Barbie says,
"Writing portable code is hard".)