Try to use smaller mmapped buffers on Ethernet devices.
Unfortunately, setting the frame size for mmapped captures based only on
the snapshot length means we end up with very large frames with the
usual large "capture the entire packet" snapshot length, which means
very *few* frames as the frames in the mmapped buffer are fixed length,
and thus perhaps more packet drops.  It's hard to choose the right frame
size in the general case, but we try clamping it at MTU+18 for Ethernet,
at least.