Use BIOCSETLIF if available.
Solaris 11 will support the ioctl BIOCSETLIF.
This operates on "struct lifname".

Whilst there are other *LIF ioctls, only this one
is important for libpcap.

For most consumers, the primary difference is the
length of the name allowed. The traditional ifname
structure only allowed for 16 character interface
names. The newer struct lifname allows for 32.
With the presence of interface renaming, amongst
other features, use of the struct lifname ioctls will
be required for proper functionality.

Reviewed-By: Guy Harris <guy@alum.mit.edu>