When creating PF_PACKET sockets, fail immediately for most errors.
If an attempt to create a PF_PACKET socket fails with anything other
than EINVAL or EAFNOSUPPORT, treat it as a fatal error; other errors
probably mean the problem isn't that you have a crufty old kernel with
no PF_PACKET support, so don't try a SOCK_PACKET socket, as it'll
probably fail, too, and cause the kernel to whine that the app is using
an obsolete API.

In all cases where we create a socket, return PCAP_ERROR_PERM_DENIED for
EPERM and EACCES, PCAP_ERROR for all other errors.

Based on a patch from Gerald Combs.