Expand comments.
Note exactly what gen_ipfrag(), and some of its callers, are doing
(i.e., rejecting packets that are the second or later fragments of a
fragmented IPv4 packet, so that you can't look at the protocol header
for the protocol running atop IPv4).

Also, note that in the IPv6 equivalents of those callers we might want
to check for the first fragment of an IPv6 packet (we currently won't
detect even the first fragment).