Fix monitor mode handling with non-mac80211 drivers.
If we don't have any private Wireless Extension ioctls to control radio
headers, we should still try to turn monitor mode off and on.

With the Wireless Extensions, you might not be able to turn monitor mode
on or off if the interface is up, so take it down first and bring it
back up afterwards.

Also add the interface name to the warning messages printed when trying
to restore modes when the program exits.