Simplify the handling of new DLT_/LINKTYPE_ values.
Add #defines for the minimum and maximum DLT_/LINKTYPE_ values in the
range of new values.

In that range, a given DLT_ and a given LINKTYPE_ have the same
numerical value; don't require entries in the DLT_-to-LINKTYPE_ table
for them.

Also, in that range, all DLT_ values are presumed valid; in the filter
code generator, for values that don't have any code to handle them,
support raw "link[N:M]" filtering rather than rejecting all filters.