Give a lot more detail, including most APIs.
Give a lot more detail, and mention most APIs, so that it's easier to
use the pcap(3PCAP) man page to find the API you want.

Both Linux and *BSD/Mac OS X man pages appear to boldface function names
but not the "this is a function" indicator parentheses after them;
follow that convention.