[PATCH] Fix AIX-3.5 crash with read failure during stress
Until libpcap.1.1.0 maximum buffer length for reads on bpf files was 32768.
(It set using  ioctl(fd, BIOCSBLEN, size) call).

In libpcap.1.1.0 this value was increased:
---
/*
 * Default capture buffer size.
 * 32K isn't very much for modern machines with fast networks; we
 * pick .5M, as that's the maximum on at least some systems with BPF.
 */
---

It does not look like a bug  - the idea is just to try this size and decrease
the value if it is too big.

However this change caused unrecoverable read failure in stress.

Probably AIX specific issue.
Probably a different (larger) value can be set.