Handle DLT_PFSYNC in {FreeBSD, other *BSD+Mac OS X, other}.
In FreeBSD, DLT_PFSYNC is defined as 121; as a link-layer type header
value in savefiles, that collides with LINKTYPE_HHDLC, so we don't want
to use it in savefiles.

In OpenBSD, NetBSD, DragonFly BSD, and Mac OS X, DLT_PFSYNC is defined
as 18; as a link-layer hader value in savefiles, that potentially
collides with DLT_CIP for ATM Classical IP in at least some versions of
SuSE, so we don't want to use it in savefiles.

In other OSes, it's not defined.

We add a new LINKTYPE_PFSYNC, with a value of 245, to use in all
savefiles, and define DLT_PFSYNC as 121 in FreeBSd, 18 in OpenBSD,
NetBSD, DragonFly BSD, and Mac OS X, and 245 elsewhere.  We map
DLT_PFSYNC <-> LINKTYPE_PFSYNC.

This means we no longer need to check whether DLT_PFSYNC is defined.