Rename some LINKTYPE_ values to make it clearer what they are.
Rename LINKTYPE_TOKEN_RING to LINKTYPE_IEEE802_5 to clarify that it
applies to IEEE 802.5 token ring, not to other token rings.

Give all the 802.11+radio metadata LINKTYPE_ values names that begin
with LINKTYPE_IEEE802_11, and rename LINKTYPE_IEEE802_11_RADIO to
LINKTYPE_IEEE802_11_RADIOTAP to clarify that it's radiotap.

At this point, the radiotap header is what's being adopted for most if
not all 802.11 adapters; get rid of the bit about the AVS header being
used for that purpose.