Don't just treat anything that contains "can" or "vcan" as a CANbus device.
Instead, require that it *start* with "can" or "vcan", and be followed
by at least one digit.

This means that if you, for example, rename a regular interface as, for
example, "canopy", it will no longer be mistakenly treated as a CANbus
device (yes, this really happened, and a bug was filed against it).