Have non-interface modules take responsibility for identifying their devices.
Have a table of routines to do pcap_create() for devices that aren't
regular network interfaces.  Try each of those in succession until one
says "it's mine" (whether it succeeds or fails); if none do, do a
pcap_create() for a regular interface.

Have those routines do more stringent tests of the name - don't just
accept any name that has a particular substring anywhere in it.  That
reduces the likelihood of a false match (as happened with the CANbus
module when somebody renamed their Ethernet interface "canopy").

Have the table also include routines for pcap_findalldevs().