Work around people who use --without-flex and --without-bison.
For some mysterious reason, some people insist on configuring libpcap
using --without-flex and --without-bison when they're running on systems
that come with Flex and Bison, *don't* offer Real Lex or Real YACC, and
thus have "lex" and "yacc" as wrappers around Flex and Bison.  At least
some versions of Bison define YYBISON regardless of whether they're run
with -y (as is done by at least some "yacc" wrappers around Bison) or
not, but the configure script only passes the -p flag to a tool named
"bison", not a tool named "yacc", so the parser ends up being built
without a #define to give the parser the name pcap_parse() and with a
YYBISON #define to suppress the pcap_parse() wrapper around yyparse(),
so you don't get a pcap_parse() defined by libpcap.

Instead, explicitly define NEED_YYPARSE_WRAPPER if we're not passing -p
to the parser generator, and use *that* to determine whether to build
the wrapper or not.

Hopefully this will put all those "ZOMG I tried to link tcpdump and it
didn't find pcap_parse()!!!!!!111!!!ONE!!!!1111" complaints to rest.