Fail if pcap_compile() handed a not-yet-activated pcap_t.
pcap_compile() needs an activated pcap_t, so that it has a link-layer
header type and snapshot length.

This should prevent people from thinking that, for example, the fix for
their program not supporting 802.11 filters when they're trying to
capture on a Wi-Fi interface on Linux or OS X but aren't capturing in
monitor mode (yes, somebody really *did* think that on stackoverflow).