Only add VLAN tags for Ethernet.
The code to insert VLAN tags was assuming Ethernet packets; that's not
the case, for example, if you're capturing on the "any" device, as the
packet headers aren't Ethernet headers.  Have a flag in the pcap
structure to indicate whether to check for and add VLAN tags, and set it
only if we're using PF_PACKET sockets, we're getting Ethernet headers,
and, if we're using the memory-mapped capture mechanism, it's not
version 1.

This should fix SourceForge bug 3525828.