Support what IPv6 stuff we can without IPv6 address resolution support.
Make the INET6-only code just be code that depends on having
getaddrinfo(), as per a complaint that, on a machine that either doesn't
have getaddrinfo() or doesn't have INET6 enabled, a filter

    dst host 127.0.0.1 and (icmp or icmp6 or ((tcp or udp or sctp) and (src
        host 127.0.0.1)))

Not having a full IPv6 stack shouldn't keep you from, for example,
having a filter that just checks the packet type field for IPv6 or that
looks at the protocol field in the IPv6 header.  You can't check for
IPv6 *hosts* (although we could perhaps allow them to be specified
numerically).