Have the capture_source_types[] array end with a terminator element.
That ensures that there's *always* at least one element in the array;
some C compilers get upset if there isn't at least one element, and if
we aren't supporting any of the additional capture types, there won't be
any elements other than the terminator.

In the loops that scan the array, look for the terminator rather than
using a count of elements.