Fix valgrind warning (jump on unitialized value) due to missing ioctl wrapper
Valgrind doesn't have a wrapper for the SIOCETHTOOL ioctl call yet. This
is causing every program that uses libpcap and valgrind to produce some
false warnings because it thinks eval.data was never initialized. A harmless
fix is to just set eval.data to 0 instead of leaving it unitialized before
the SIOCETHTOOL ioctl call sets it.