Cope with Bluetooth API breakage in some Linux 3.x(?) kernel.
In the 2.6.32.4 kernel, struct sockaddr_hci has no hci_channel member.

In the 3.0.4 kernel, struct sockaddr_hci has an hci_channel member, and
it has to get set in order for binds of AF_BLUETOOTH/BTPROTO_HCI sockets
to work.

Cope with this API breakage by checking whether struct sockaddr_hci has
an hci_channel member and, if it does, set it to HCI_CHANNEL_RAW.