Rename pcap_findalldevs() to pcap_findalldevs_interfaces().
pcap_findalldevs() is now a libpcap API, and expects the main "find all
interfaces" call for the platform to be named
pcap_findalldevs_interfaces(); make it so.

Also, note that the first PacketGetAdapterNames() is trying to find out
how big the packet buffer should be, and should always return FALSE.
Initialize the NameLength variable, because PacketGetAdapterNames() may
refer to it before it's set (harmlessly, but might as well be clean
here, at least until it's fixed to check the pointer argument first).