Add a test program for the open, close, and filter compile/set code paths.
This is to test for both libpcap bugs *and* Valgrind bugs; Valgrind on
OS X (and probably other systems using BPF) and Linux reports bogus
errors when setting a filter (because it blindly checks that the
*entire* bpf_program structure for BIOCSETF and the *entire* sock_fprog
structure for setsockopt(SOL_SOCKET, SO_ATTACH_FILTER) has been
initialized, rather than checking only the structure members and not the
padding, and *doesn't* check whether the program *pointed to* by that
structure has been initialized).  See

	https://bugs.kde.org/show_bug.cgi?id=312989

(I'll be filing a similar bug, with a similar patch to fix it, for
Valgrind-on-Linux).