Pick up some fixes from a comment for the SourceForge bug for D-Bus support.
Fix the help message for --enable-dbus to say --enable-dbus, not
--enable-can, and call pcap_cleanup_live_common() rather than the
nonexistent pcap_live_cleanup_common().

Also, say "D-Bus capture" in the help message, to spell it correctly and
to clarify that we're only adding support for capturing on D-Bus (not
adding support for libpcap using D-Bus for some purpose or another).