Fix D-Bus checks.
Check whether dbus-1 is installed before trying to get flags for it.
That way, if you don't have dbus-1 installed, you don't get pkg-config
errors printed.

Don't expect the shell to be able to run a command inside $() and have
the result expand to the standard output of that command; that's not a
feature present in all Bourne-compatible shells.  Use backquotes
instead.

If the user specified --enable-dbus, and dbus-1 wasn't found, fail -
they asked for D-Bus sniffing support, but we can't give it to them, so
let them know that they can't have what they want.  They either need to
install the package or stop specifying --enable-dbus.