Move platform-dependent pcap_t data out of the pcap_t structure.
Put the private data right after the pcap_t structure, with a pointer to
it in the pcap_t.

The initial goal is to allow new pcap modules to be added without having
to hack pcap-int.h.

In the longer term, we may want to freeze the pcap_t structure, except
possibly for adding new method pointers at the end, and provide an ABI
for adding modules.

We also put the stuff used by the read path at the beginning of the
pcap_t structure, to try to keep it on the same set of cache lines.