Don't try -W flags with HP's C compiler.
HP's C compiler cannot be made to exit with a non-zero exit status when
given an unknown compiler flag, so you can't use that to figure out if a
flag is supported, but it *will* print a warning, which we don't want.

In addition, it doesn't support -W flags for warnings in any case, so
let's not bother with them.