Fix handling of time stamps in pcap-ng files.
There's the time stamp resolution for a particular interface in a file,
and there's the time stamp resolution the user asked for; the former is
what it is, regardless of the time stamp resolution the user asked for.

Add support for multiple interfaces with different time stamps while
we're at it.